feat: move options filter to common folder
