+++ public/components/common/index.ts
@@ -9,3 +9,4 @@ export * from './tag_filter_popover_content';
 export * from './selected_tag_filter_panel';
 export * from './debounced_search_bar';
 export * from './tag_filter';
+export * from './options_filter';
