+++ public/components/common/options_filter/__tests__/options_filter_item.test.tsx
@@ -6,16 +6,16 @@
 import React from 'react';
 import userEvent from '@testing-library/user-event';
 
-import { ModelFilterItem } from '../model_filter_item';
+import { OptionsFilterItem } from '../options_filter_item';
 
-import { render, screen } from '../../../../test/test_utils';
+import { render, screen } from '../../../../../test/test_utils';
 
-describe('<ModelFilterItem />', () => {
+describe('<OptionsFilterItem />', () => {
   it('should render passed children and check icon', () => {
     render(
-      <ModelFilterItem checked="on" value="foo" onClick={() => {}}>
+      <OptionsFilterItem checked="on" value="foo" onClick={() => {}}>
         foo
-      </ModelFilterItem>
+      </OptionsFilterItem>
     );
     expect(screen.getByText('foo')).toBeInTheDocument();
     expect(screen.getByRole('img', { hidden: true })).toBeInTheDocument();
@@ -24,9 +24,9 @@ describe('<ModelFilterItem />', () => {
   it('should call onClick with "foo" after click', async () => {
     const onClickMock = jest.fn();
     render(
-      <ModelFilterItem checked="on" value="foo" onClick={onClickMock}>
+      <OptionsFilterItem checked="on" value="foo" onClick={onClickMock}>
         foo
-      </ModelFilterItem>
+      </OptionsFilterItem>
     );
     await userEvent.click(screen.getByRole('option'));
     expect(onClickMock).toHaveBeenCalledWith('foo');
