+++ public/components/common/options_filter/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { OptionsFilter, OptionsFilterProps } from './options_filter';
