+++ public/components/common/options_filter/options_filter_item.tsx
@@ -6,13 +6,18 @@
 import React, { useCallback } from 'react';
 import { EuiFilterSelectItem, EuiFilterSelectItemProps } from '@elastic/eui';
 
-export interface ModelFilterItemProps
+export interface OptionsFilterItemProps
   extends Pick<EuiFilterSelectItemProps, 'checked' | 'children'> {
   value: string;
   onClick: (value: string) => void;
 }
 
-export const ModelFilterItem = ({ checked, children, onClick, value }: ModelFilterItemProps) => {
+export const OptionsFilterItem = ({
+  checked,
+  children,
+  onClick,
+  value,
+}: OptionsFilterItemProps) => {
   const handleClick = useCallback(() => {
     onClick(value);
   }, [onClick, value]);
